Moved record field name enum over to the projections package.
