+++ adam-commands/src/main/scala/edu/berkeley/cs/amplab/adam/projections/Projection.scala
@@ -3,7 +3,6 @@ package edu.berkeley.cs.amplab.adam.projections
 import org.apache.avro.Schema
 import scala.collection.JavaConversions._
 import org.apache.avro.Schema.Field
-import edu.berkeley.cs.amplab.adam.predicates.ADAMRecordField
 import edu.berkeley.cs.amplab.adam.avro.ADAMRecord
 
 /**
